requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
colorama==0.4.6
Django==4.1.4
django-taggit==3.0.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
